isolation of mealplanning stuff
