+++ backend/cmd/localdev/server/meal_planning_bootstrap.go
@@ -1,0 +1,513 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"maps"
+	"slices"
+	"time"
+
+	mealplangrocerylistinitializerbuild "github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/build/jobs/meal_plan_grocery_list_initializer"
+	mealplantaskcreatorbuild "github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/build/jobs/meal_plan_task_creator"
+	"github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/config"
+	"github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/domain/mealplanning"
+	"github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/domain/mealplanning/bootstrap"
+	"github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/domain/mealplanning/managers"
+	"github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/domain/mealplanning/recipeanalysis"
+	mealplanningrepo "github.com/dinnerdonebetter/dinnerdonebetter/backend/internal/repositories/postgres/mealplanning"
+
+	"github.com/verygoodsoftwarenotvirus/platform/v4/identifiers"
+	msgconfig "github.com/verygoodsoftwarenotvirus/platform/v4/messagequeue/config"
+	noopmq "github.com/verygoodsoftwarenotvirus/platform/v4/messagequeue/noop"
+	"github.com/verygoodsoftwarenotvirus/platform/v4/observability/logging"
+	"github.com/verygoodsoftwarenotvirus/platform/v4/observability/metrics"
+	"github.com/verygoodsoftwarenotvirus/platform/v4/observability/tracing"
+	textsearchcfg "github.com/verygoodsoftwarenotvirus/platform/v4/search/text/config"
+)
+
+var createMealPlansAndVotes = true // os.Getenv("CREATE_MEAL_PLANS_AND_VOTES") == "true"
+
+func bootstrapEnumerationsAndRecipes(ctx context.Context, repo mealplanning.Repository, logger logging.Logger, tracerProvider tracing.TracerProvider, adminUserID string) error {
+	logger.Info("Creating enumerations...")
+	enums, enumsErr := bootstrap.CreateEnumerations(ctx, repo, logger)
+	if enumsErr != nil {
+		return fmt.Errorf("failed to create enumerations: %w", enumsErr)
+	}
+	logger.Info("Enumerations created successfully!")
+
+	// Create RecipeManager to create the first recipe
+	logger.Info("Creating RecipeManager...")
+	queueCfg := &msgconfig.QueuesConfig{
+		DataChangesTopicName: "data_changes",
+	}
+	publisherProvider := noopmq.NewPublisherProvider()
+	recipeAnalyzer := recipeanalysis.NewRecipeAnalyzer(logger, tracerProvider)
+	searchConfig := &textsearchcfg.Config{}
+	metricsProvider := metrics.NewNoopMetricsProvider()
+
+	recipeManager, recipeManagerErr := managers.NewRecipeManager(
+		ctx,
+		logger,
+		tracerProvider,
+		repo,
+		queueCfg,
+		publisherProvider,
+		recipeAnalyzer,
+		searchConfig,
+		metricsProvider,
+	)
+	if recipeManagerErr != nil {
+		return fmt.Errorf("failed to create recipe manager: %w", recipeManagerErr)
+	}
+	logger.Info("RecipeManager created successfully!")
+
+	logger.Info("Creating remaining bootstrap recipes...")
+
+	// Phase 1: Create recipes without prerequisites
+	allRecipes := bootstrap.AllRecipes(enums)
+	logger.Info(fmt.Sprintf("Found %d recipes without prerequisites to create", len(allRecipes)))
+
+	createdRecipes := make(map[string]*mealplanning.Recipe)
+	// Create recipes without prerequisites
+	for i, recipe := range allRecipes {
+		logger.Info(fmt.Sprintf("Creating recipe %d: %s (%d steps)", i+1, recipe.Name, len(recipe.Steps)))
+		r, createErr := recipeManager.CreateRecipe(ctx, adminUserID, recipe)
+		if createErr != nil {
+			return fmt.Errorf("failed to create recipe #%d %s: %w", i, recipe.Name, createErr)
+		}
+
+		createdRecipes[r.Slug] = r
+	}
+	logger.Info("All recipes without prerequisites created successfully!")
+
+	recipes := slices.Collect(maps.Values(createdRecipes))
+
+	// Phase 2: Create recipes with prerequisites
+	recipesWithPrerequisites := bootstrap.AllRecipesWithPrerequisites(enums, createdRecipes)
+	logger.Info(fmt.Sprintf("Found %d recipes with prerequisites to create", len(recipesWithPrerequisites)))
+
+	for i, recipe := range recipesWithPrerequisites {
+		// Resolve empty recipe IDs in cross-recipe references before creating
+		// This is needed because getRecipeIDBySlug may return empty strings when
+		// called during recipe input construction (before prerequisite recipes exist)
+		resolveEmptyRecipeIDs(recipe, createdRecipes)
+
+		logger.Info(fmt.Sprintf("Creating recipe with prerequisites %d: %s (%d steps)", i+1, recipe.Name, len(recipe.Steps)))
+		r, createErr := recipeManager.CreateRecipe(ctx, adminUserID, recipe)
+		if createErr != nil {
+			return fmt.Errorf("failed to create recipe with prerequisites #%d %s: %w", i, recipe.Name, createErr)
+		}
+
+		recipes = append(recipes, r)
+		// Update lookup map so subsequent recipes in phase 2 can reference this one
+		createdRecipes[r.Slug] = r
+	}
+	logger.Info("All bootstrap recipes created successfully!")
+
+	// Approve all bootstrap recipes
+	for _, r := range recipes {
+		if err := recipeManager.UpdateRecipeStatus(ctx, r.ID, mealplanning.RecipeStatusApproved); err != nil {
+			return fmt.Errorf("failed to approve recipe %s: %w", r.Name, err)
+		}
+	}
+	logger.Info("All bootstrap recipes approved!")
+
+	// Always create meals
+	logger.Info("Creating bootstrap meals...")
+	meals := bootstrap.AllMeals(adminUserID, recipes)
+	logger.Info(fmt.Sprintf("Found %d meals to create", len(meals)))
+
+	for i, meal := range meals {
+		logger.Info(fmt.Sprintf("Creating meal %d: %s (%d components)", i+1, meal.Name, len(meal.Components)))
+		if _, err := repo.CreateMeal(ctx, meal); err != nil {
+			return fmt.Errorf("failed to create meal %s: %w", meal.Name, err)
+		}
+	}
+	logger.Info("All bootstrap meals created successfully!")
+
+	return nil
+}
+
+func bootstrapMealPlan(ctx context.Context, repo mealplanning.Repository, logger logging.Logger, adminUserID, adminAccountID string) (string, error) {
+	if !createMealPlansAndVotes {
+		logger.Info("Skipping meal plan creation (CREATE_MEAL_PLANS_AND_VOTES=false)")
+		return "", nil
+	}
+	if adminUserID == "" || adminAccountID == "" {
+		return "", fmt.Errorf("admin user ID or account ID not set")
+	}
+
+	logger.Info("Creating meal plan with chicken dishes...")
+
+	// Get all meals created by admin user
+	mealsResult, mealsErr := repo.GetMealsCreatedByUser(ctx, adminUserID, nil)
+	if mealsErr != nil {
+		return "", fmt.Errorf("failed to get meals: %w", mealsErr)
+	}
+
+	// Find the 3 chicken dishes
+	chickenMealNames := []string{
+		"Sous Vide Chicken Breast with Rice",
+		"Roast Chicken with Caesar Broccoli",
+		"Soy Sauce Braised Chicken Thighs with Rice",
+	}
+
+	otherMealNames := []string{
+		"Pan-Seared Steak with Mashed Potatoes",
+		"Classic Burgers with Mixed Green Salad",
+		"Grilled Pork Tenderloin with Brussels Sprouts",
+	}
+
+	var (
+		chickenMeals []*mealplanning.Meal
+		otherMeals   []*mealplanning.Meal
+	)
+	for _, meal := range mealsResult.Data {
+		if slices.Contains(chickenMealNames, meal.Name) {
+			chickenMeals = append(chickenMeals, meal)
+		}
+
+		for _, name := range otherMealNames {
+			if meal.Name == name {
+				otherMeals = append(otherMeals, meal)
+			}
+		}
+	}
+
+	now := time.Now()
+
+	// Voting deadline is Friday before the event (midnight)
+	votingDeadline := cloneTime(now).Add(24 * time.Hour * 3)
+
+	eventStart := cloneTime(votingDeadline).Add(24 * time.Hour)
+	eventEnd := cloneTime(eventStart).Add(2 * time.Hour) // 2 hour duration
+
+	// Create options for all three chicken meals
+	var chickenOptions []*mealplanning.MealPlanOptionDatabaseCreationInput
+	for _, chickenMeal := range chickenMeals {
+		chickenOptions = append(chickenOptions, &mealplanning.MealPlanOptionDatabaseCreationInput{
+			ID:        identifiers.New(),
+			MealID:    chickenMeal.ID,
+			MealScale: 1.0,
+		})
+	}
+
+	var otherOptions []*mealplanning.MealPlanOptionDatabaseCreationInput
+	for _, otherMeal := range otherMeals {
+		otherOptions = append(otherOptions, &mealplanning.MealPlanOptionDatabaseCreationInput{
+			ID:        identifiers.New(),
+			MealID:    otherMeal.ID,
+			MealScale: 1.0,
+		})
+	}
+
+	// Create a single event with all three chickenOptions
+	events := []*mealplanning.MealPlanEventDatabaseCreationInput{
+		{
+			ID:       identifiers.New(),
+			StartsAt: eventStart,
+			EndsAt:   eventEnd,
+			MealName: mealplanning.DinnerMealName,
+			Options:  chickenOptions,
+		},
+		{
+			ID:       identifiers.New(),
+			StartsAt: cloneTime(eventStart).Add(24 * time.Hour),
+			EndsAt:   cloneTime(eventEnd).Add(24 * time.Hour),
+			MealName: mealplanning.SupperMealName,
+			Options:  otherOptions,
+		},
+	}
+
+	// Create meal plan
+	mealPlanInput := &mealplanning.MealPlanDatabaseCreationInput{
+		ID:               identifiers.New(),
+		Notes:            "Example \tMeal Plan",
+		VotingDeadline:   votingDeadline,
+		ElectionMethod:   mealplanning.MealPlanElectionMethodSchulze,
+		BelongsToAccount: adminAccountID,
+		CreatedByUser:    adminUserID,
+		Events:           events,
+	}
+
+	createdMealPlan, mealPlanErr := repo.CreateMealPlan(ctx, mealPlanInput)
+	if mealPlanErr != nil {
+		return "", fmt.Errorf("failed to create meal plan: %w", mealPlanErr)
+	}
+
+	logger.Info(fmt.Sprintf("Created meal plan %s with %d events", createdMealPlan.ID, len(events)))
+	return createdMealPlan.ID, nil
+}
+
+func bootstrapFinalizedMealPlanAndVotes(ctx context.Context, repo mealplanning.Repository, logger logging.Logger, adminUserID, adminAccountID, currentMealPlanID string, memberUserIDs []string) error {
+	if !createMealPlansAndVotes {
+		logger.Info("Skipping finalized meal plan and vote creation (CREATE_MEAL_PLANS_AND_VOTES=false)")
+		return nil
+	}
+	if adminUserID == "" || adminAccountID == "" {
+		return fmt.Errorf("admin user ID or account ID not set")
+	}
+
+	if currentMealPlanID == "" {
+		return fmt.Errorf("current meal plan ID not set")
+	}
+
+	if len(memberUserIDs) == 0 {
+		return fmt.Errorf("member user IDs not set")
+	}
+
+	// Get the current meal plan to use its timing
+	currentMealPlan, mealPlanErr := repo.GetMealPlan(ctx, currentMealPlanID, adminAccountID)
+	if mealPlanErr != nil {
+		return fmt.Errorf("failed to get current meal plan: %w", mealPlanErr)
+	}
+
+	logger.Info("Creating finalized meal plan with votes...")
+
+	// Get all meals created by admin user
+	mealsResult, mealsErr := repo.GetMealsCreatedByUser(ctx, adminUserID, nil)
+	if mealsErr != nil {
+		return fmt.Errorf("failed to get meals: %w", mealsErr)
+	}
+
+	// Find the 3 chicken dishes
+	chickenMealNames := []string{
+		"Sous Vide Chicken Breast with Rice",
+		"Roast Chicken with Caesar Broccoli",
+		"Soy Sauce Braised Chicken Thighs with Rice",
+	}
+
+	otherMealNames := []string{
+		"Pan-Seared Steak with Mashed Potatoes",
+		"Classic Burgers with Mixed Green Salad",
+		"Grilled Pork Tenderloin with Brussels Sprouts",
+	}
+
+	var (
+		chickenMeals []*mealplanning.Meal
+		otherMeals   []*mealplanning.Meal
+	)
+	for _, meal := range mealsResult.Data {
+		if slices.Contains(chickenMealNames, meal.Name) {
+			chickenMeals = append(chickenMeals, meal)
+		}
+
+		for _, name := range otherMealNames {
+			if meal.Name == name {
+				otherMeals = append(otherMeals, meal)
+			}
+		}
+	}
+
+	// Use the same timing as the current meal plan
+	finalizedVotingDeadline := currentMealPlan.VotingDeadline
+	finalizedEventStart := cloneTime(finalizedVotingDeadline).Add(24 * time.Hour)
+	finalizedEventEnd := cloneTime(finalizedEventStart).Add(2 * time.Hour)
+
+	// Create options for all three chicken meals
+	var chickenOptions []*mealplanning.MealPlanOptionDatabaseCreationInput
+	for _, chickenMeal := range chickenMeals {
+		chickenOptions = append(chickenOptions, &mealplanning.MealPlanOptionDatabaseCreationInput{
+			ID:        identifiers.New(),
+			MealID:    chickenMeal.ID,
+			MealScale: 1.0,
+		})
+	}
+
+	var otherOptions []*mealplanning.MealPlanOptionDatabaseCreationInput
+	for _, otherMeal := range otherMeals {
+		otherOptions = append(otherOptions, &mealplanning.MealPlanOptionDatabaseCreationInput{
+			ID:        identifiers.New(),
+			MealID:    otherMeal.ID,
+			MealScale: 1.0,
+		})
+	}
+
+	// Create events for finalized meal plan
+	finalizedEvents := []*mealplanning.MealPlanEventDatabaseCreationInput{
+		{
+			ID:       identifiers.New(),
+			StartsAt: finalizedEventStart,
+			EndsAt:   finalizedEventEnd,
+			MealName: mealplanning.DinnerMealName,
+			Options:  chickenOptions,
+		},
+		{
+			ID:       identifiers.New(),
+			StartsAt: cloneTime(finalizedEventStart).Add(24 * time.Hour),
+			EndsAt:   cloneTime(finalizedEventEnd).Add(24 * time.Hour),
+			MealName: mealplanning.SupperMealName,
+			Options:  otherOptions,
+		},
+	}
+
+	// Create finalized meal plan
+	finalizedMealPlanInput := &mealplanning.MealPlanDatabaseCreationInput{
+		ID:               identifiers.New(),
+		Notes:            "Finalized Example Meal Plan",
+		VotingDeadline:   finalizedVotingDeadline,
+		ElectionMethod:   mealplanning.MealPlanElectionMethodSchulze,
+		BelongsToAccount: adminAccountID,
+		CreatedByUser:    adminUserID,
+		Events:           finalizedEvents,
+	}
+
+	finalizedMealPlan, finalizedErr := repo.CreateMealPlan(ctx, finalizedMealPlanInput)
+	if finalizedErr != nil {
+		return fmt.Errorf("failed to create finalized meal plan: %w", finalizedErr)
+	}
+
+	logger.Info(fmt.Sprintf("Created finalized meal plan %s with %d events", finalizedMealPlan.ID, len(finalizedEvents)))
+
+	// Create votes from all members for all options in all events
+	// We need to reload the meal plan to get the created options
+	finalizedMealPlanWithEvents, finalizeErr := repo.GetMealPlan(ctx, finalizedMealPlan.ID, adminAccountID)
+	if finalizeErr != nil {
+		return fmt.Errorf("failed to get finalized meal plan with events: %w", finalizeErr)
+	}
+
+	for _, event := range finalizedMealPlanWithEvents.Events {
+		for _, memberUserID := range memberUserIDs {
+			// Create votes for this user for all options in this event
+			var votes []*mealplanning.MealPlanOptionVoteDatabaseCreationInput
+			for rank, option := range event.Options {
+				votes = append(votes, &mealplanning.MealPlanOptionVoteDatabaseCreationInput{
+					ID:                      identifiers.New(),
+					ByUser:                  memberUserID,
+					BelongsToMealPlanOption: option.ID,
+					Rank:                    uint8(rank),
+					Abstain:                 false,
+					Notes:                   "",
+				})
+			}
+
+			// Create votes for this user
+			voteInput := &mealplanning.MealPlanOptionVotesDatabaseCreationInput{
+				Votes: votes,
+			}
+			if _, voteErr := repo.CreateMealPlanOptionVote(ctx, voteInput); voteErr != nil {
+				return fmt.Errorf("failed to create votes for user %s: %w", memberUserID, voteErr)
+			}
+		}
+	}
+
+	logger.Info("Created votes from all members for finalized meal plan")
+
+	// Finalize the meal plan (idempotent: already-finalized is OK on re-run)
+	finalized, finalizeErr := repo.AttemptToFinalizeMealPlan(ctx, finalizedMealPlan.ID, adminAccountID)
+	switch {
+	case finalizeErr != nil:
+		if errors.Is(finalizeErr, mealplanningrepo.ErrAlreadyFinalized) {
+			logger.Info("Meal plan already finalized (idempotent re-run), continuing")
+		} else {
+			return fmt.Errorf("failed to finalize meal plan: %w", finalizeErr)
+		}
+	case finalized:
+		logger.Info("Finalized meal plan successfully")
+	default:
+		return fmt.Errorf("meal plan was not finalized")
+	}
+
+	// Extend the current meal plan's voting deadline by one week
+	updatedMealPlan := *currentMealPlan
+	updatedMealPlan.VotingDeadline = cloneTime(currentMealPlan.VotingDeadline).Add(7 * 24 * time.Hour)
+	if updateErr := repo.UpdateMealPlan(ctx, &updatedMealPlan); updateErr != nil {
+		return fmt.Errorf("failed to update current meal plan voting deadline: %w", updateErr)
+	}
+
+	logger.Info(fmt.Sprintf("Extended current meal plan %s voting deadline by one week", currentMealPlanID))
+	return nil
+}
+
+func bootstrapMealPlanWorkers(ctx context.Context, logger logging.Logger, apiConfig *config.APIServiceConfig) error {
+	if !createMealPlansAndVotes {
+		logger.Info("Skipping grocery list and task creator workers (CREATE_MEAL_PLANS_AND_VOTES=false)")
+		return nil
+	}
+	logger.Info("Running grocery list initializer and task creator workers...")
+
+	// Build grocery list initializer worker
+	groceryListConfig := &config.MealPlanGroceryListInitializerConfig{
+		Database:      apiConfig.Database,
+		Observability: apiConfig.Observability,
+		Events:        apiConfig.Events,
+		Queues:        apiConfig.Queues,
+		Analytics:     apiConfig.Analytics,
+	}
+	groceryListWorker, workerErr := mealplangrocerylistinitializerbuild.Build(ctx, groceryListConfig)
+	if workerErr != nil {
+		return fmt.Errorf("failed to build grocery list initializer worker: %w", workerErr)
+	}
+
+	// Run grocery list initializer
+	if err := groceryListWorker.Work(ctx); err != nil {
+		return fmt.Errorf("failed to run grocery list initializer worker: %w", err)
+	}
+	logger.Info("Grocery list initializer worker completed successfully")
+
+	// Build task creator worker
+	taskCreatorConfig := &config.MealPlanTaskCreatorConfig{
+		Database:      apiConfig.Database,
+		Observability: apiConfig.Observability,
+		Events:        apiConfig.Events,
+		Queues:        apiConfig.Queues,
+		Analytics:     apiConfig.Analytics,
+	}
+	taskCreatorWorker, workerErr := mealplantaskcreatorbuild.Build(ctx, taskCreatorConfig)
+	if workerErr != nil {
+		return fmt.Errorf("failed to build task creator worker: %w", workerErr)
+	}
+
+	// Run task creator
+	if err := taskCreatorWorker.Work(ctx); err != nil {
+		return fmt.Errorf("failed to run task creator worker: %w", err)
+	}
+	logger.Info("Task creator worker completed successfully")
+
+	return nil
+}
+
+func cloneTime(t time.Time) time.Time {
+	t, parseErr := time.Parse(time.RFC3339, t.Format(time.RFC3339))
+	if parseErr != nil {
+		panic(parseErr)
+	}
+
+	return t
+}
+
+// resolveEmptyRecipeIDs resolves empty RecipeStepProductRecipeID values in a recipe input
+// by looking up recipes in the createdRecipes map. Prefers RecipeStepProductRecipeSlug when
+// set (ensures correct recipe); otherwise falls back to matching by step index (ambiguous).
+func resolveEmptyRecipeIDs(recipe *mealplanning.RecipeCreationRequestInput, createdRecipes map[string]*mealplanning.Recipe) {
+	for _, step := range recipe.Steps {
+		for _, ingredient := range step.Ingredients {
+			if ingredient.RecipeStepProductRecipeID != nil && *ingredient.RecipeStepProductRecipeID == "" {
+				// Prefer slug lookup when available - ensures we resolve to the correct recipe
+				if ingredient.RecipeStepProductRecipeSlug != nil && *ingredient.RecipeStepProductRecipeSlug != "" {
+					if refRecipe, ok := createdRecipes[*ingredient.RecipeStepProductRecipeSlug]; ok && refRecipe != nil {
+						ingredient.RecipeStepProductRecipeID = &refRecipe.ID
+						continue
+					}
+				}
+				// Fallback: match by step index (ambiguous - many recipes may have a step at that index)
+				stepIndex := ingredient.ProductOfRecipeStepIndex
+				if stepIndex != nil {
+					for _, refRecipe := range createdRecipes {
+						if refRecipe != nil && int(*stepIndex) < len(refRecipe.Steps) {
+							referencedStep := refRecipe.Steps[*stepIndex]
+							if ingredient.ProductOfRecipeStepProductIndex != nil {
+								productIndex := int(*ingredient.ProductOfRecipeStepProductIndex)
+								if productIndex < len(referencedStep.Products) {
+									ingredient.RecipeStepProductRecipeID = &refRecipe.ID
+									break
+								}
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+}
