+++ frontend/packages/api-client/src/comments/comments_service_types.ts
@@ -12,36 +12,6 @@ import { Comment, CommentCreationRequestInput, CommentUpdateRequestInput } from 
 
 export const protobufPackage = 'comments';
 
-export interface AddCommentToMealPlanRequest {
-  mealPlanId: string;
-  input: CommentCreationRequestInput | undefined;
-}
-
-export interface AddCommentToMealPlanResponse {
-  responseDetails: ResponseDetails | undefined;
-  comment: Comment | undefined;
-}
-
-export interface AddCommentToMealRequest {
-  mealId: string;
-  input: CommentCreationRequestInput | undefined;
-}
-
-export interface AddCommentToMealResponse {
-  responseDetails: ResponseDetails | undefined;
-  comment: Comment | undefined;
-}
-
-export interface AddCommentToRecipeRequest {
-  recipeId: string;
-  input: CommentCreationRequestInput | undefined;
-}
-
-export interface AddCommentToRecipeResponse {
-  responseDetails: ResponseDetails | undefined;
-  comment: Comment | undefined;
-}
-
 export interface CreateCommentRequest {
   input: CommentCreationRequestInput | undefined;
 }
@@ -80,507 +50,6 @@ export interface UpdateCommentResponse {
   responseDetails: ResponseDetails | undefined;
   comment: Comment | undefined;
 }
-
-function createBaseAddCommentToMealPlanRequest(): AddCommentToMealPlanRequest {
-  return { mealPlanId: '', input: undefined };
-}
-
-export const AddCommentToMealPlanRequest: MessageFns<AddCommentToMealPlanRequest> = {
-  encode(message: AddCommentToMealPlanRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.mealPlanId !== '') {
-      writer.uint32(10).string(message.mealPlanId);
-    }
-    if (message.input !== undefined) {
-      CommentCreationRequestInput.encode(message.input, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToMealPlanRequest {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToMealPlanRequest();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.mealPlanId = reader.string();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.input = CommentCreationRequestInput.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToMealPlanRequest {
-    return {
-      mealPlanId: isSet(object.mealPlanId)
-        ? globalThis.String(object.mealPlanId)
-        : isSet(object.meal_plan_id)
-          ? globalThis.String(object.meal_plan_id)
-          : '',
-      input: isSet(object.input) ? CommentCreationRequestInput.fromJSON(object.input) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToMealPlanRequest): unknown {
-    const obj: any = {};
-    if (message.mealPlanId !== '') {
-      obj.mealPlanId = message.mealPlanId;
-    }
-    if (message.input !== undefined) {
-      obj.input = CommentCreationRequestInput.toJSON(message.input);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToMealPlanRequest>, I>>(base?: I): AddCommentToMealPlanRequest {
-    return AddCommentToMealPlanRequest.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToMealPlanRequest>, I>>(object: I): AddCommentToMealPlanRequest {
-    const message = createBaseAddCommentToMealPlanRequest();
-    message.mealPlanId = object.mealPlanId ?? '';
-    message.input =
-      object.input !== undefined && object.input !== null
-        ? CommentCreationRequestInput.fromPartial(object.input)
-        : undefined;
-    return message;
-  },
-};
-
-function createBaseAddCommentToMealPlanResponse(): AddCommentToMealPlanResponse {
-  return { responseDetails: undefined, comment: undefined };
-}
-
-export const AddCommentToMealPlanResponse: MessageFns<AddCommentToMealPlanResponse> = {
-  encode(message: AddCommentToMealPlanResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.responseDetails !== undefined) {
-      ResponseDetails.encode(message.responseDetails, writer.uint32(10).fork()).join();
-    }
-    if (message.comment !== undefined) {
-      Comment.encode(message.comment, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToMealPlanResponse {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToMealPlanResponse();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.responseDetails = ResponseDetails.decode(reader, reader.uint32());
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.comment = Comment.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToMealPlanResponse {
-    return {
-      responseDetails: isSet(object.responseDetails)
-        ? ResponseDetails.fromJSON(object.responseDetails)
-        : isSet(object.response_details)
-          ? ResponseDetails.fromJSON(object.response_details)
-          : undefined,
-      comment: isSet(object.comment) ? Comment.fromJSON(object.comment) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToMealPlanResponse): unknown {
-    const obj: any = {};
-    if (message.responseDetails !== undefined) {
-      obj.responseDetails = ResponseDetails.toJSON(message.responseDetails);
-    }
-    if (message.comment !== undefined) {
-      obj.comment = Comment.toJSON(message.comment);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToMealPlanResponse>, I>>(base?: I): AddCommentToMealPlanResponse {
-    return AddCommentToMealPlanResponse.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToMealPlanResponse>, I>>(object: I): AddCommentToMealPlanResponse {
-    const message = createBaseAddCommentToMealPlanResponse();
-    message.responseDetails =
-      object.responseDetails !== undefined && object.responseDetails !== null
-        ? ResponseDetails.fromPartial(object.responseDetails)
-        : undefined;
-    message.comment =
-      object.comment !== undefined && object.comment !== null ? Comment.fromPartial(object.comment) : undefined;
-    return message;
-  },
-};
-
-function createBaseAddCommentToMealRequest(): AddCommentToMealRequest {
-  return { mealId: '', input: undefined };
-}
-
-export const AddCommentToMealRequest: MessageFns<AddCommentToMealRequest> = {
-  encode(message: AddCommentToMealRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.mealId !== '') {
-      writer.uint32(10).string(message.mealId);
-    }
-    if (message.input !== undefined) {
-      CommentCreationRequestInput.encode(message.input, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToMealRequest {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToMealRequest();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.mealId = reader.string();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.input = CommentCreationRequestInput.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToMealRequest {
-    return {
-      mealId: isSet(object.mealId)
-        ? globalThis.String(object.mealId)
-        : isSet(object.meal_id)
-          ? globalThis.String(object.meal_id)
-          : '',
-      input: isSet(object.input) ? CommentCreationRequestInput.fromJSON(object.input) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToMealRequest): unknown {
-    const obj: any = {};
-    if (message.mealId !== '') {
-      obj.mealId = message.mealId;
-    }
-    if (message.input !== undefined) {
-      obj.input = CommentCreationRequestInput.toJSON(message.input);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToMealRequest>, I>>(base?: I): AddCommentToMealRequest {
-    return AddCommentToMealRequest.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToMealRequest>, I>>(object: I): AddCommentToMealRequest {
-    const message = createBaseAddCommentToMealRequest();
-    message.mealId = object.mealId ?? '';
-    message.input =
-      object.input !== undefined && object.input !== null
-        ? CommentCreationRequestInput.fromPartial(object.input)
-        : undefined;
-    return message;
-  },
-};
-
-function createBaseAddCommentToMealResponse(): AddCommentToMealResponse {
-  return { responseDetails: undefined, comment: undefined };
-}
-
-export const AddCommentToMealResponse: MessageFns<AddCommentToMealResponse> = {
-  encode(message: AddCommentToMealResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.responseDetails !== undefined) {
-      ResponseDetails.encode(message.responseDetails, writer.uint32(10).fork()).join();
-    }
-    if (message.comment !== undefined) {
-      Comment.encode(message.comment, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToMealResponse {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToMealResponse();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.responseDetails = ResponseDetails.decode(reader, reader.uint32());
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.comment = Comment.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToMealResponse {
-    return {
-      responseDetails: isSet(object.responseDetails)
-        ? ResponseDetails.fromJSON(object.responseDetails)
-        : isSet(object.response_details)
-          ? ResponseDetails.fromJSON(object.response_details)
-          : undefined,
-      comment: isSet(object.comment) ? Comment.fromJSON(object.comment) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToMealResponse): unknown {
-    const obj: any = {};
-    if (message.responseDetails !== undefined) {
-      obj.responseDetails = ResponseDetails.toJSON(message.responseDetails);
-    }
-    if (message.comment !== undefined) {
-      obj.comment = Comment.toJSON(message.comment);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToMealResponse>, I>>(base?: I): AddCommentToMealResponse {
-    return AddCommentToMealResponse.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToMealResponse>, I>>(object: I): AddCommentToMealResponse {
-    const message = createBaseAddCommentToMealResponse();
-    message.responseDetails =
-      object.responseDetails !== undefined && object.responseDetails !== null
-        ? ResponseDetails.fromPartial(object.responseDetails)
-        : undefined;
-    message.comment =
-      object.comment !== undefined && object.comment !== null ? Comment.fromPartial(object.comment) : undefined;
-    return message;
-  },
-};
-
-function createBaseAddCommentToRecipeRequest(): AddCommentToRecipeRequest {
-  return { recipeId: '', input: undefined };
-}
-
-export const AddCommentToRecipeRequest: MessageFns<AddCommentToRecipeRequest> = {
-  encode(message: AddCommentToRecipeRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.recipeId !== '') {
-      writer.uint32(10).string(message.recipeId);
-    }
-    if (message.input !== undefined) {
-      CommentCreationRequestInput.encode(message.input, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToRecipeRequest {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToRecipeRequest();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.recipeId = reader.string();
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.input = CommentCreationRequestInput.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToRecipeRequest {
-    return {
-      recipeId: isSet(object.recipeId)
-        ? globalThis.String(object.recipeId)
-        : isSet(object.recipe_id)
-          ? globalThis.String(object.recipe_id)
-          : '',
-      input: isSet(object.input) ? CommentCreationRequestInput.fromJSON(object.input) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToRecipeRequest): unknown {
-    const obj: any = {};
-    if (message.recipeId !== '') {
-      obj.recipeId = message.recipeId;
-    }
-    if (message.input !== undefined) {
-      obj.input = CommentCreationRequestInput.toJSON(message.input);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToRecipeRequest>, I>>(base?: I): AddCommentToRecipeRequest {
-    return AddCommentToRecipeRequest.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToRecipeRequest>, I>>(object: I): AddCommentToRecipeRequest {
-    const message = createBaseAddCommentToRecipeRequest();
-    message.recipeId = object.recipeId ?? '';
-    message.input =
-      object.input !== undefined && object.input !== null
-        ? CommentCreationRequestInput.fromPartial(object.input)
-        : undefined;
-    return message;
-  },
-};
-
-function createBaseAddCommentToRecipeResponse(): AddCommentToRecipeResponse {
-  return { responseDetails: undefined, comment: undefined };
-}
-
-export const AddCommentToRecipeResponse: MessageFns<AddCommentToRecipeResponse> = {
-  encode(message: AddCommentToRecipeResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.responseDetails !== undefined) {
-      ResponseDetails.encode(message.responseDetails, writer.uint32(10).fork()).join();
-    }
-    if (message.comment !== undefined) {
-      Comment.encode(message.comment, writer.uint32(18).fork()).join();
-    }
-    return writer;
-  },
-
-  decode(input: BinaryReader | Uint8Array, length?: number): AddCommentToRecipeResponse {
-    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
-    const end = length === undefined ? reader.len : reader.pos + length;
-    const message = createBaseAddCommentToRecipeResponse();
-    while (reader.pos < end) {
-      const tag = reader.uint32();
-      switch (tag >>> 3) {
-        case 1: {
-          if (tag !== 10) {
-            break;
-          }
-
-          message.responseDetails = ResponseDetails.decode(reader, reader.uint32());
-          continue;
-        }
-        case 2: {
-          if (tag !== 18) {
-            break;
-          }
-
-          message.comment = Comment.decode(reader, reader.uint32());
-          continue;
-        }
-      }
-      if ((tag & 7) === 4 || tag === 0) {
-        break;
-      }
-      reader.skip(tag & 7);
-    }
-    return message;
-  },
-
-  fromJSON(object: any): AddCommentToRecipeResponse {
-    return {
-      responseDetails: isSet(object.responseDetails)
-        ? ResponseDetails.fromJSON(object.responseDetails)
-        : isSet(object.response_details)
-          ? ResponseDetails.fromJSON(object.response_details)
-          : undefined,
-      comment: isSet(object.comment) ? Comment.fromJSON(object.comment) : undefined,
-    };
-  },
-
-  toJSON(message: AddCommentToRecipeResponse): unknown {
-    const obj: any = {};
-    if (message.responseDetails !== undefined) {
-      obj.responseDetails = ResponseDetails.toJSON(message.responseDetails);
-    }
-    if (message.comment !== undefined) {
-      obj.comment = Comment.toJSON(message.comment);
-    }
-    return obj;
-  },
-
-  create<I extends Exact<DeepPartial<AddCommentToRecipeResponse>, I>>(base?: I): AddCommentToRecipeResponse {
-    return AddCommentToRecipeResponse.fromPartial(base ?? ({} as any));
-  },
-  fromPartial<I extends Exact<DeepPartial<AddCommentToRecipeResponse>, I>>(object: I): AddCommentToRecipeResponse {
-    const message = createBaseAddCommentToRecipeResponse();
-    message.responseDetails =
-      object.responseDetails !== undefined && object.responseDetails !== null
-        ? ResponseDetails.fromPartial(object.responseDetails)
-        : undefined;
-    message.comment =
-      object.comment !== undefined && object.comment !== null ? Comment.fromPartial(object.comment) : undefined;
-    return message;
-  },
-};
 
 function createBaseCreateCommentRequest(): CreateCommentRequest {
   return { input: undefined };
