+++ proto/mealplanning/mealplanning_service_types.proto
@@ -2,6 +2,7 @@ syntax = "proto3";
 
 package mealplanning;
 
+import "comments/comments_messages.proto";
 import "common.proto";
 import "filtering.proto";
 import "google/protobuf/timestamp.proto";
@@ -3134,4 +3135,34 @@ message UploadRecipeStepImageRequest {
 message UploadRecipeStepImageResponse {
   common.ResponseDetails response_details = 1;
   optional string uploaded_media_id = 2;
+}
+
+message AddCommentToMealPlanRequest {
+  string meal_plan_id = 1;
+  comments.CommentCreationRequestInput input = 2;
+}
+
+message AddCommentToMealPlanResponse {
+  common.ResponseDetails response_details = 1;
+  comments.Comment comment = 2;
+}
+
+message AddCommentToMealRequest {
+  string meal_id = 1;
+  comments.CommentCreationRequestInput input = 2;
+}
+
+message AddCommentToMealResponse {
+  common.ResponseDetails response_details = 1;
+  comments.Comment comment = 2;
+}
+
+message AddCommentToRecipeRequest {
+  string recipe_id = 1;
+  comments.CommentCreationRequestInput input = 2;
+}
+
+message AddCommentToRecipeResponse {
+  common.ResponseDetails response_details = 1;
+  comments.Comment comment = 2;
 }
